+++ tests/firedrake/test_block_system.py
@@ -1,4 +1,11 @@
-from .test_base import setup_test  # noqa: F401
+from firedrake import *
+from tlm_adjoint.firedrake import *
+from tlm_adjoint.firedrake.backend_interface import matrix_multiply
+from tlm_adjoint.firedrake.block_system import (
+    BlockMatrix, BlockNullspace, ConstantNullspace, DirichletBCNullspace,
+    LinearSolver as _BlockLinearSolver, UnityNullspace)
+
+from .test_base import *
 
 import mpi4py.MPI as MPI  # noqa: N817
 import numpy as np
@@ -10,16 +17,6 @@ pytestmark = pytest.mark.skipif(
     MPI.COMM_WORLD.size not in {1, 4},
     reason="tests must be run in serial, or with 4 processes")
 
-from tlm_adjoint.firedrake.block_system import (  # noqa: E402
-    BlockMatrix, BlockNullspace, ConstantNullspace, DirichletBCNullspace,
-    LinearSolver as _BlockLinearSolver, UnityNullspace)
-
-from firedrake import (  # noqa: E402
-    Constant, ConvergenceError, DirichletBC, Function, FunctionSpace,
-    LinearSolver, SpatialCoordinate, TestFunction, TrialFunction,
-    UnitSquareMesh, VectorFunctionSpace, action, adjoint, assemble, ds, dx,
-    exp, grad, inner, pi, sin, solve)
-
 
 class BlockLinearSolver(_BlockLinearSolver):
     def solve(self, *args, **kwargs):
@@ -30,7 +27,8 @@ class BlockLinearSolver(_BlockLinearSolver):
 
 @pytest.mark.firedrake
 @pytest.mark.parametrize("pc", ["none", "block_jacobi", "block_chebyshev"])
-def test_block_diagonal(setup_test, pc):  # noqa: F811
+@seed_test
+def test_block_diagonal(setup_test, pc):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = FunctionSpace(mesh, "Lagrange", 1)
@@ -146,7 +144,8 @@ def test_block_diagonal(setup_test, pc):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_constant_nullspace(setup_test):  # noqa: F811
+@seed_test
+def test_constant_nullspace(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = FunctionSpace(mesh, "Lagrange", 1)
@@ -197,7 +196,8 @@ def test_constant_nullspace(setup_test):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_unity_nullspace(setup_test):  # noqa: F811
+@seed_test
+def test_unity_nullspace(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = FunctionSpace(mesh, "Lagrange", 1)
@@ -247,7 +247,8 @@ def test_unity_nullspace(setup_test):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_dirichlet_bc_nullspace(setup_test):  # noqa: F811
+@seed_test
+def test_dirichlet_bc_nullspace(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = FunctionSpace(mesh, "Lagrange", 1)
@@ -297,7 +298,8 @@ def test_dirichlet_bc_nullspace(setup_test):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_pressure_projection(setup_test):  # noqa: F811
+@seed_test
+def test_pressure_projection(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = VectorFunctionSpace(mesh, "Lagrange", 2)
@@ -378,7 +380,8 @@ def test_pressure_projection(setup_test):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_mass(setup_test):  # noqa: F811
+@seed_test
+def test_mass(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space = FunctionSpace(mesh, "Lagrange", 1)
@@ -405,7 +408,8 @@ def test_mass(setup_test):  # noqa: F811
 
 
 @pytest.mark.firedrake
-def test_sub_block(setup_test):  # noqa: F811
+@seed_test
+def test_sub_block(setup_test):
     mesh = UnitSquareMesh(10, 10)
     X = SpatialCoordinate(mesh)
     space_0 = FunctionSpace(mesh, "Lagrange", 1)
@@ -480,3 +484,85 @@ def test_sub_block(setup_test):  # noqa: F811
     u_2_error_norm = np.sqrt(abs(assemble(inner(u_2 - u_2_ref,
                                                 u_2 - u_2_ref) * dx)))
     assert u_2_error_norm < 1.0e-12
+
+
+@pytest.mark.firedrake
+@no_space_type_checking
+@seed_test
+def test_HEP(setup_test, test_leaks):
+    mesh = UnitIntervalMesh(20)
+    space = FunctionSpace(mesh, "Lagrange", 1)
+    test, trial = TestFunction(space), TrialFunction(space)
+
+    M = assemble(inner(trial, test) * dx)
+
+    def M_action(x, y):
+        assemble(inner(x, test) * dx, tensor=y)
+
+    Lam, V = eigensolve(space, space.dual(), M_action,
+                        solver_parameters={"eps_type": "krylovschur",
+                                           "eps_hermitian": None,
+                                           "eps_largest_magnitude": None,
+                                           "eps_nev": space.dim(),
+                                           "eps_conv_rel": None,
+                                           "eps_tol": 1.0e-12,
+                                           "eps_purify": False})
+
+    assert issubclass(Lam.dtype.type, np.floating)
+    assert (Lam > 0.0).all()
+
+    error = Function(space)
+    assert len(Lam) == len(V)
+    for lam, (v_r, v_i) in zip(Lam, V):
+        assert abs(var_inner(v_r, v_r) - 1.0) < 1.0e-14
+        matrix_multiply(M, v_r, tensor=error)
+        var_axpy(error, -lam, v_r)
+        assert var_linf_norm(error) < 1.0e-16
+        assert v_i is None
+
+
+@pytest.mark.firedrake
+@no_space_type_checking
+@seed_test
+def test_NHEP(setup_test, test_leaks):
+    mesh = UnitIntervalMesh(20)
+    space = FunctionSpace(mesh, "Lagrange", 1)
+    test, trial = TestFunction(space), TrialFunction(space)
+
+    N = assemble(inner(trial.dx(0), test) * dx)
+
+    def N_action(x, y):
+        assemble(inner(x.dx(0), test) * dx, tensor=y)
+
+    Lam, V = eigensolve(space, space.dual(), N_action,
+                        solver_parameters={"eps_type": "krylovschur",
+                                           "eps_non_hermitian": None,
+                                           "eps_largest_magnitude": None,
+                                           "eps_nev": space.dim(),
+                                           "eps_conv_rel": None,
+                                           "eps_tol": 1.0e-12,
+                                           "eps_purify": False})
+
+    assert issubclass(Lam.dtype.type, np.complexfloating)
+    assert abs(Lam.real).max() < 1.0e-14
+
+    error = Function(space)
+    assert len(Lam) == len(V)
+    if issubclass(PETSc.ScalarType, np.floating):
+        for lam, (v_r, v_i) in zip(Lam, V):
+            assert abs(var_inner(v_r, v_r) + var_inner(v_i, v_i) - 1.0) < 1.0e-14  # noqa: E501
+            matrix_multiply(N, v_r, tensor=error)
+            var_axpy(error, -lam.real, v_r)
+            var_axpy(error, +lam.imag, v_i)
+            assert var_linf_norm(error) < 1.0e-15
+            matrix_multiply(N, v_i, tensor=error)
+            var_axpy(error, -lam.real, v_i)
+            var_axpy(error, -lam.imag, v_r)
+            assert var_linf_norm(error) < 1.0e-15
+    else:
+        for lam, (v_r, v_i) in zip(Lam, V):
+            assert abs(var_inner(v_r, v_r) - 1.0) < 1.0e-14
+            matrix_multiply(N, v_r, tensor=error)
+            var_axpy(error, -lam, v_r)
+            assert var_linf_norm(error) < 1.0e-14
+            assert v_i is None
