+++ tests/firedrake/test_eigendecomposition.py
@@ -1,11 +1,9 @@
 from firedrake import *
 from tlm_adjoint.firedrake import *
-from tlm_adjoint.firedrake.backend_interface import matrix_multiply
 
 from .test_base import *
 
 import numpy as np
-import petsc4py.PETSc as PETSc
 import pytest
 try:
     import slepc4py.SLEPc as SLEPc
@@ -18,88 +16,6 @@ pytestmark = pytest.mark.skipif(
 pytestmark = pytest.mark.skipif(
     SLEPc is None,
     reason="SLEPc not available")
-
-
-@pytest.mark.firedrake
-@no_space_type_checking
-@seed_test
-def test_HEP(setup_test, test_leaks):
-    mesh = UnitIntervalMesh(20)
-    space = FunctionSpace(mesh, "Lagrange", 1)
-    test, trial = TestFunction(space), TrialFunction(space)
-
-    M = assemble(inner(trial, test) * dx)
-
-    def M_action(x, y):
-        assemble(inner(x, test) * dx, tensor=y)
-
-    Lam, V = eigensolve(space, space.dual(), M_action,
-                        solver_parameters={"eps_type": "krylovschur",
-                                           "eps_hermitian": None,
-                                           "eps_largest_magnitude": None,
-                                           "eps_nev": space.dim(),
-                                           "eps_conv_rel": None,
-                                           "eps_tol": 1.0e-12,
-                                           "eps_purify": False})
-
-    assert issubclass(Lam.dtype.type, np.floating)
-    assert (Lam > 0.0).all()
-
-    error = Function(space)
-    assert len(Lam) == len(V)
-    for lam, (v_r, v_i) in zip(Lam, V):
-        assert abs(var_inner(v_r, v_r) - 1.0) < 1.0e-14
-        matrix_multiply(M, v_r, tensor=error)
-        var_axpy(error, -lam, v_r)
-        assert var_linf_norm(error) < 1.0e-16
-        assert v_i is None
-
-
-@pytest.mark.firedrake
-@no_space_type_checking
-@seed_test
-def test_NHEP(setup_test, test_leaks):
-    mesh = UnitIntervalMesh(20)
-    space = FunctionSpace(mesh, "Lagrange", 1)
-    test, trial = TestFunction(space), TrialFunction(space)
-
-    N = assemble(inner(trial.dx(0), test) * dx)
-
-    def N_action(x, y):
-        assemble(inner(x.dx(0), test) * dx, tensor=y)
-
-    Lam, V = eigensolve(space, space.dual(), N_action,
-                        solver_parameters={"eps_type": "krylovschur",
-                                           "eps_non_hermitian": None,
-                                           "eps_largest_magnitude": None,
-                                           "eps_nev": space.dim(),
-                                           "eps_conv_rel": None,
-                                           "eps_tol": 1.0e-12,
-                                           "eps_purify": False})
-
-    assert issubclass(Lam.dtype.type, np.complexfloating)
-    assert abs(Lam.real).max() < 1.0e-14
-
-    error = Function(space)
-    assert len(Lam) == len(V)
-    if issubclass(PETSc.ScalarType, np.floating):
-        for lam, (v_r, v_i) in zip(Lam, V):
-            assert abs(var_inner(v_r, v_r) + var_inner(v_i, v_i) - 1.0) < 1.0e-14  # noqa: E501
-            matrix_multiply(N, v_r, tensor=error)
-            var_axpy(error, -lam.real, v_r)
-            var_axpy(error, +lam.imag, v_i)
-            assert var_linf_norm(error) < 1.0e-15
-            matrix_multiply(N, v_i, tensor=error)
-            var_axpy(error, -lam.real, v_i)
-            var_axpy(error, -lam.imag, v_r)
-            assert var_linf_norm(error) < 1.0e-15
-    else:
-        for lam, (v_r, v_i) in zip(Lam, V):
-            assert abs(var_inner(v_r, v_r) - 1.0) < 1.0e-14
-            matrix_multiply(N, v_r, tensor=error)
-            var_axpy(error, -lam, v_r)
-            assert var_linf_norm(error) < 1.0e-14
-            assert v_i is None
 
 
 @pytest.mark.firedrake
